requirements.txt
streamlit
anthropic
langdetect
python-dotenv